+++ packages/camera/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.3.3
+* Appaly new texture APIs.
+
 ## 0.3.2
 
 * Update the example app and integration_test.