+++ packages/video_player/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 2.4.1
+* Appaly new texture APIs.
+
 ## 2.4.0
 
 * Update video_player to 2.4.2.