+++ packages/webview_flutter/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.5.3
+* Appaly new texture APIs.
+
 ## 0.5.2
 * Add a back key handling.